Pull out `:room-consumer` module
