+++ core/src/main/kotlin/com/zynger/floorplan/dbml/Field.kt
@@ -1,8 +1,8 @@
 package com.zynger.floorplan.dbml
 
 import com.zynger.floorplan.Settings
-import com.zynger.floorplan.model.PrimaryKey
-import com.zynger.floorplan.model.Field as DbField
+import com.zynger.floorplan.room.PrimaryKey
+import com.zynger.floorplan.room.Field as DbField
 
 class Field(
     private val field: DbField,
