+++ core/src/main/kotlin/com/zynger/floorplan/dbml/Index.kt
@@ -1,7 +1,7 @@
 package com.zynger.floorplan.dbml
 
 import java.lang.StringBuilder
-import com.zynger.floorplan.model.Index as DbIndex
+import com.zynger.floorplan.room.Index as DbIndex
 
 class Index(
     private val index: DbIndex
