+++ core/src/main/kotlin/com/zynger/floorplan/dbml/Reference.kt
@@ -1,7 +1,7 @@
 package com.zynger.floorplan.dbml
 
-import com.zynger.floorplan.model.ForeignKey
-import com.zynger.floorplan.model.ForeignKeyAction
+import com.zynger.floorplan.room.ForeignKey
+import com.zynger.floorplan.room.ForeignKeyAction
 import java.lang.StringBuilder
 
 class Reference(
