+++ floorplan-cli/src/main/kotlin/com/zynger/floorplan/App.kt
@@ -1,20 +1,17 @@
 package com.zynger.floorplan
 
 import com.zynger.floorplan.dbml.Table
-import com.zynger.floorplan.model.Schema
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonConfiguration
+import com.zynger.floorplan.room.Database
+import com.zynger.floorplan.room.RoomConsumer
 import java.io.File
 
 fun main(args: Array<String>) {
     val input = InputParser.parse(args)
     val src = File(input.schemaPath)
-    val json = Json(JsonConfiguration.Stable.copy(ignoreUnknownKeys = true, isLenient = true))
+    val database: Database = RoomConsumer.read(src)
 
     val settings = Settings(input.creationSqlAsTableNote, input.renderNullableFields)
-    val dbml = json
-        .parse(Schema.serializer(), src.readText())
-        .database
+    val dbml = database
         .entities
         .map { Table(it, settings) }
         .joinToString(separator = "\n\n")
