+++ room-consumer/build.gradle
@@ -1,11 +1,8 @@
 plugins {
-    id 'application'
     id 'java'
     id 'org.jetbrains.kotlin.jvm'
+    id 'org.jetbrains.kotlin.plugin.serialization' version '1.3.71'
 }
-
-group 'zynger'
-version '0.1-SNAPSHOT'
 
 sourceCompatibility = 1.8
 
@@ -14,14 +11,9 @@ repositories {
     jcenter()
 }
 
-application {
-    mainClassName = "com.zynger.floorplan.AppKt"
-}
-
 dependencies {
-    implementation project(":core")
-    implementation project(":room-consumer")
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+    implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.20.0" // JVM dependency
     testImplementation "junit:junit:4.12"
 }
