+++ settings.gradle
@@ -2,3 +2,4 @@ rootProject.name = 'floorplan'
 include 'core'
 include 'dbml-parser'
 include 'floorplan-cli'
+include 'room-consumer'
